Full Site JS Rewrite
The site had a bunch of bugs, and a few of naive algorithms. Algorithms were rewritten to have a slightly stronger statistical basis, and more proper data parsing was instituted. I'd still like the final scaled score algorithm to be a function of a normal distribution or something similar, but didn't quite have the time to implement. I tried to stay faithful to the structure of the original algorithm, just with more reliable measuring methods as parameters. 

Partial Changelog since I forgot everything that was changed:
- More proper ACT to SAT conversion based upon the CollegeBoard's concordance tables.
- Rewrite of data acquisition to assure proper parsing.
- Unobfuscation of a few variables, as well as increased commenting.
- Changed SAT scaling in objective score to be based off of percentile, not ratio. This should create more accurate distributions.
- Changed scaled score algorithm since after correct data parsing it would return a number 90 or higher most of the time. The new algorithm was sorta thrown together and I want to change it too.

I love the work done on this site, just thought there was some improvements to make so I made them.